Realized that addStylist.js was a redundant file. Moved and renamed the component associated with Stylist Registratt Registration to StylistSignUp.js
