+++ hair/src/components/StylistSignUp.js
@@ -1,0 +1,73 @@
+import React, {useState} from 'react';
+
+//path for localhost: http://localhost:3000/addstylist
+
+export default function StylistSignUp(props) {
+    const {setStylist} = props;
+    const [newStylist, setNewStylist] = useState({name:" ", number:" ", email:" "})
+
+    //change event"
+    function handleChange(event) {
+        const updatedStylist = {...newStylist, [event.target.name]: event.target.value};
+        console.log(
+            "handleChange",
+            event.target.name,
+            event.target.value,
+            updatedStylist
+        );
+        setNewStylist(updatedStylist);
+    }
+
+    //submit event
+    function handleSubmit(event) {
+        event.preventDefault();
+        setStylist(person => [...person, newStylist])
+        // submitStylist(newStylist)
+        setNewStylist({name:" ", number:" ", email:" "})
+
+        console.log('name', newStylist);
+    }
+
+    return (
+        <form onSubmit={handleSubmit}>
+            {/* Name */}
+            <div className="form-group">
+                <label>Name:</label>
+                <input
+                    type="text"
+                    className="input"
+                    name="name"
+                    placeholder="Please enter your name"
+                    value={newStylist.name}
+                    onChange={handleChange}
+                />
+            </div>
+            {/* Phone Number */}
+            <div className="form-group">
+                <label>Phone Number:</label>
+                <input
+                    
+                    className="input"
+                    name="number"
+                    placeholder="Please enter your number"
+                    value={newStylist.number}
+                    onChange={handleChange}
+                />
+            </div>
+            {/* Email */}
+            <div className="form-group">
+                <label>Email:</label>
+                <input
+                    type="email"
+                    className="input"
+                    name="email"
+                    placeholder="Please enter your email"
+                    value={newStylist.email}
+                    onChange={handleChange}
+                />
+            </div>
+            <button type="submit" className="add-button">Add</button>
+        </form>
+    )
+
+};
